+++ graylog2-web-interface/public/stylesheets/bootstrap-config.json
@@ -19,7 +19,7 @@
     "@link-hover-decoration": "underline",
     "@font-family-sans-serif": "\"Open Sans\", \"Helvetica Neue\", Helvetica, Arial, sans-serif",
     "@font-family-serif": "Georgia, \"Times New Roman\", Times, serif",
-    "@font-family-monospace": "Menlo, Monaco, Consolas, \"Courier New\", monospace",
+    "@font-family-monospace": "\"Roboto Mono\", Menlo, Monaco, Consolas, \"Courier New\", monospace",
     "@font-family-base": "@font-family-sans-serif",
     "@font-size-base": "14px",
     "@font-size-large": "ceil((@font-size-base * 1.25))",
