+++ graylog2-web-interface/src/theme/GraylogThemeProvider.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { ThemeProvider } from 'styled-components';
 
-import { color, utils } from 'theme';
+import { colors, fonts, utils } from 'theme';
 import buttonStyles from 'components/graylog/styles/buttonStyles';
 
 /* NOTE: mode will eventually need to come from User Preferences */
@@ -12,9 +12,10 @@ const GraylogThemeProvider = ({ children }) => {
   return (
     <ThemeProvider theme={{
       mode: THEME_MODE,
-      color: color[THEME_MODE],
+      color: colors[THEME_MODE],
+      fonts,
       components: {
-        button: buttonStyles({ color: color[THEME_MODE] }),
+        button: buttonStyles({ color: colors[THEME_MODE] }),
       },
       utils,
     }}>
