+++ graylog2-web-interface/src/theme/fonts.js
@@ -1,0 +1,24 @@
+// @flow strict
+
+import 'opensans-npm-webfont/open_sans.css';
+import 'opensans-npm-webfont/open_sans_italic.css';
+import 'opensans-npm-webfont/open_sans_bold.css';
+import '@openfonts/roboto-mono_latin/index.css';
+
+export type Fonts = {
+  family: {
+    body: string,
+    monospace: string,
+  },
+};
+
+const family = {
+  body: '"Open Sans", "Helvetica Neue", Helvetica, Arial, sans-serif',
+  monospace: '"Roboto Mono", Menlo, Monaco, Consolas, "Courier New", monospace',
+};
+
+const fonts: Fonts = {
+  family,
+};
+
+export default fonts;
